+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This section is for maintaining a changelog for all breaking changes for the cli
 
 ### Added
 - Document HTTP/2 support ([#330](https://github.com/opensearch-project/opensearch-java/pull/330))
+- Added support for "withJson" in PutIndexTemplateRequest ([#657](https://github.com/opensearch-project/opensearch-java/pull/657))
+
 
 ### Dependencies